requirements.txt
moviepy
Pillow
pydub
pyrogram
python-decouple
rarfile
TgCrypto
unrar